+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/passport/compare/v12.0.2...12.x)
+## [Unreleased](https://github.com/laravel/passport/compare/v12.2.0...12.x)
+
+## [v12.2.0](https://github.com/laravel/passport/compare/v12.1.0...v12.2.0) - 2024-04-17
+
+* [12.x] Add `refreshToken` relation to `Token` model by [@gdebrauwer](https://github.com/gdebrauwer) in https://github.com/laravel/passport/pull/1739
+
+## [v12.1.0](https://github.com/laravel/passport/compare/v12.0.3...v12.1.0) - 2024-04-15
+
+* [12.x] Make Passport's database connection configurable by [@axlon](https://github.com/axlon) in https://github.com/laravel/passport/pull/1738
+
+## [v12.0.3](https://github.com/laravel/passport/compare/v12.0.2...v12.0.3) - 2024-04-05
+
+* Adjust newFactory method visibility by [@brandonfarber](https://github.com/brandonfarber) in https://github.com/laravel/passport/pull/1735
 
 ## [v12.0.2](https://github.com/laravel/passport/compare/v12.0.1...v12.0.2) - 2024-03-21